fix: address Copilot and Devin review findings
- Move envBanner.ts from src/bootstrap/ to src/utils/ so the
  infrastructure CLI (richmenu/cli.ts) no longer imports from bootstrap,
  fixing the inverted dependency direction flagged by Copilot.
- Banner now prints both ENV and NODE_ENV so production cases with an
  unset ENV no longer render a misleading 'ENV=LOCAL' value.
- Update the banner JSDoc to reflect that LOCAL_DEV-gated scripts
  include richmenu:deploy variants, not just pnpm dev*.
- Propagate the LOCAL_DEV guard into presentation/graphql/server.ts so
  ApolloServerPluginUsageReporting stays disabled when running locally
  against dev/prd (Devin review follow-up).
- Add an isTTY check to scripts/confirmPrd.ts; the prd confirmation
  prompt now aborts instead of hanging when stdin is not a TTY.
- Reformat reservation/config.ts to 2-space indentation to match the
  repository Prettier config.
