+++ src/infrastructure/richmenu/cli.ts
@@ -6,7 +6,7 @@ import { getCommunityConfig, getAllCommunityIds } from "./index";
 import { deployRichMenus } from "./deployer";
 import { DeployRichMenuContext } from "./types";
 import logger from "@/infrastructure/logging";
-import { printEnvBanner } from "@/bootstrap/envBanner";
+import { printEnvBanner } from "@/utils/envBanner";
 
 printEnvBanner();
